fix(pr-review): move AnaphoraPredicate to frontend.datalog and use PatternWalker base for eq mixin
Addresses PR review comments:
- AnaphoraPredicate is SQUALL-specific, not a general logic operator, so it belongs in frontend.datalog (moved to anaphora_resolution.py)
- ConditionAwareEqMixin should explicitly inherit PatternWalker as a pattern-matching mixin

Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)
