+++ neurolang/logic/__init__.py
@@ -149,29 +149,6 @@ class ExistentialPredicate(Quantifier):
         return r
 
 
-class AnaphoraPredicate(ExistentialPredicate):
-    """An existential introduced by 'the X' when X was not in symbol scope.
-
-    The noun_name (a Symbol) records which noun the anaphora refers to,
-    enabling a post-processing walker to resolve it against the matching
-    noun predicate on the other side of a Condition boundary.
-    """
-    def __init__(self, head, body, noun_name):
-        super().__init__(head, body)
-        if not isinstance(noun_name, Symbol):
-            raise NeuroLangException(
-                'A Symbol should be provided for the noun name of the '
-                'anaphora expression'
-            )
-        self.noun_name = noun_name
-
-    def __repr__(self):
-        r = (
-            u'\u2203{{{} st {} | anaphora={}}}'
-            .format(self.head, self.body, self.noun_name)
-        )
-        return r
-
 
 class UniversalPredicate(Quantifier):
     def __init__(self, head, body):
